Move results calculation to results app
