+++ evap/results/templates/result_bar.html
@@ -1,4 +1,4 @@
-{% load evaluation_filters %}
+{% load results_templatetags %}
 
 {% if show_grades and result.total_count > 0 %}
     {% spaceless %}
