+++ evap/results/templates/results_course_detail.html
@@ -1,6 +1,6 @@
 {% extends "results_base.html" %}
 
-{% load evaluation_filters %}
+{% load results_templatetags %}
 
 {% block title %}{{ course.name }} - {{ course.semester.name}} - {{ block.super }}{% endblock %}
