+++ evap/results/templates/results_semester_detail.html
@@ -1,5 +1,6 @@
 {% extends "results_base.html" %}
 
+{% load results_templatetags %}
 {% load evaluation_filters %}
 
 {% block title %}{{ semester.name }} - {{ block.super }}{% endblock %}
