+++ evap/results/views.py
@@ -5,7 +5,7 @@ from django.shortcuts import get_object_or_404, render
 from django.contrib.auth.decorators import login_required
 
 from evap.evaluation.models import Semester, Degree, Contribution
-from evap.evaluation.tools import calculate_results, calculate_average_grades_and_deviation, TextResult, RatingResult
+from evap.results.tools import calculate_results, calculate_average_grades_and_deviation, TextResult, RatingResult
 
 
 @login_required
